requirements.txt
google-api-python-client
google-auth-oauthlib
telebot
google-api-core
googleapis-common-protos
httplib2